+++ lib/ethers.ex
@@ -142,6 +142,7 @@ defmodule Ethers do
   ## Parameters
   - tx_hash: Transaction hash which the transaction is queried for.
   - overrides:
+    - rpc_client: The RPC module to use for this request (overrides default).
     - rpc_opts: Specific RPC options to specify for this request.
   """
   @spec get_transaction(Types.t_hash(), Keyword.t()) ::
@@ -151,6 +152,24 @@ defmodule Ethers do
 
     rpc_client.eth_get_transaction_by_hash(tx_hash, rpc_opts)
     |> post_process(nil, :get_transaction)
+  end
+
+  @doc """
+  Returns the receipt of a transaction by it's hash.
+
+  ## Parameters
+  - tx_hash: Transaction hash which the transaction receipt is queried for.
+  - overrides:
+    - rpc_client: The RPC module to use for this request (overrides default).
+    - rpc_opts: Specific RPC options to specify for this request.
+  """
+  @spec get_transaction_receipt(Types.t_hash(), Keyword.t()) ::
+          {:ok, map()} | {:error, term()}
+  def get_transaction_receipt(tx_hash, opts \\ []) when is_binary(tx_hash) do
+    {rpc_client, rpc_opts} = get_rpc_client(opts)
+
+    rpc_client.eth_get_transaction_receipt(tx_hash, rpc_opts)
+    |> post_process(nil, :get_transaction_receipt)
   end
 
   @doc """
@@ -644,6 +663,9 @@ defmodule Ethers do
 
   defp post_process({:ok, nil}, _tx_hash, :get_transaction),
     do: {:error, :transaction_not_found}
+
+  defp post_process({:ok, nil}, _tx_hash, :get_transaction_receipt),
+    do: {:error, :transaction_receipt_not_found}
 
   defp post_process({:ok, result}, _tx_data, _action),
     do: {:ok, result}